refactor: 예외 패키지 생성 후 패키지 이동
