+++ jdbc/src/main/java/com/interface21/jdbc/core/JdbcTemplate.java
@@ -12,6 +12,8 @@ import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.interface21.jdbc.exception.SqlExecutionException;
+
 public class JdbcTemplate {
 
     private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
