+++ jdbc/src/main/java/com/interface21/jdbc/exception/ConnectionCloseException.java
@@ -1,10 +1,6 @@
-package com.interface21.jdbc.core;
+package com.interface21.jdbc.exception;
 
 public class ConnectionCloseException extends RuntimeException {
-
-    public ConnectionCloseException(String message) {
-        super(message);
-    }
 
     public ConnectionCloseException(final String message, final Throwable cause) {
         super(message, cause);
