+++ jdbc/src/main/java/com/interface21/jdbc/exception/TransactionRollbackException.java
@@ -1,10 +1,6 @@
-package com.interface21.jdbc.core;
+package com.interface21.jdbc.exception;
 
 public class TransactionRollbackException extends RuntimeException {
-
-    public TransactionRollbackException(String message) {
-        super(message);
-    }
 
     public TransactionRollbackException(final String message, final Throwable cause) {
         super(message, cause);
