+++ jdbc/src/test/java/com/interface21/jdbc/core/TransactionManagerTest.java
@@ -14,6 +14,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import com.interface21.jdbc.exception.SqlExecutionException;
+import com.interface21.jdbc.exception.TransactionExecutionException;
+
 class TransactionManagerTest {
 
     private Connection connection;
